podman/requirements/base.txt
# Base dependencies for all services
numpy>=1.21.0
pandas>=1.3.0
python-dateutil>=2.8.0
pytz>=2021.1
pyyaml>=5.4.0

# Database and ORM
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
alembic>=1.11.0
cassandra-driver>=3.28.0

# Computer Vision Core
opencv-python>=4.5.0
torch>=1.13.0
torchvision>=0.14.0
pillow>=8.3.0

# OCR and Face Recognition
face-recognition>=1.3.0
dlib>=19.24.0
pytesseract>=0.3.10
easyocr>=1.6.0

# Logging and Monitoring
prometheus-client>=0.14.0
python-json-logger>=2.0.0
loguru>=0.6.0

# Utilities
tqdm>=4.64.0
click>=8.0.0
rich>=12.0.0
requests>=2.25.0
scikit-learn>=1.0.0 